requirements.txt
ultralytics
openvino>=2024.0.0
fastapi
uvicorn
opencv-python-headless
numpy<2.0.0
python-dotenv